pyproject.toml
[project]
name = "food-nutrition-ai"
version = "0.1.0"
description = "이미지 기반 음식 인식 및 영양 정보 제공 AI 시스템 (LangChain + Gemini + SQLite)"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0,<3.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
ingest  = "src.ingest:main"
analyze = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]